% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.apropos.R
\name{maxima.apropos}
\alias{maxima.apropos}
\title{Uses maximas lookup function \code{apropos()} and retrieves result}
\usage{
maxima.apropos(keystring)
}
\arguments{
\item{command}{A character vector containing the maxima command.}
}
\value{
A character vector containing the corresponding output from maxima
}
\description{
Uses maximas lookup function \code{apropos()} and retrieves result
}
\examples{
maxima.start()

maxima.apropos("int")
}
