% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.engine.R
\name{maxima.engine}
\alias{maxima.engine}
\title{knitr maxima engine}
\usage{
maxima.engine(options)
}
\arguments{
\item{options}{A list of chunk options}
}
\value{
This functions prints the resulting output from maxima together with it's code
}
\description{
An R-function that is registered as a knitr engine when package \code{rmaxima} is attached, i.e. \code{library(rmaxima)}.
}
\details{
\code{maxima.engine} is called by \code{knit()} to evaluate maxima code chunks. When called upon the first code chunk of a document it spawns a child process running Maxima in the background. This means that a single Maxima session is used for all Maxima code chunks of an RMarkdown document. 

In addition, this function sets up Maxima specific output and chunk hooks.
}
