% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.load.R
\name{maxima.load}
\alias{maxima.load}
\title{Let maxima load a module}
\usage{
maxima.load(module)
}
\arguments{
\item{module}{A character vector naming the maxima module to be loaded.}
}
\value{
A character vector containing the corresponding output from maxima
}
\description{
Let maxima load a module
}
\examples{
maxima.load("ctensor")
}
