% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{maxima.options}
\alias{maxima.options}
\title{maxima.options}
\usage{
maxima.options(
  ...,
  RESET = FALSE,
  READ.ONLY = NULL,
  LOCAL = FALSE,
  ADD = FALSE
)
}
\arguments{
\item{format}{Character vector of length 1 setting the output format for \code{maxima.get()}. Can be one of "linear", "ascii", "latex" or "mathml".}

\item{engine.format}{Same as option \code{format}, but for outputs in RMarkdown documents.}

\item{inline.format}{Character vector of length 1 setting the output format for \code{maxima.inline()}, for knitting outputs inline into RMarkdown documents. Can be one of "linear", "latex" or "mathml". Note: Not "ascii".}

\item{label}{Logical of length 1, whether output reference labels should be printed for returned S3 objects from \code{maxima.get()} (TRUE, default), or not (FALSE).}

\item{engine.label}{Same as \code{label}, but for outputs in RMarkdown documents.}

\item{inline.label}{Same as \code{label}, but for inline outputs in RMarkdown documents.}
}
\description{
Function for globally setting and retrieving options.
}
\details{
To print out a table of available options, current settings and a description simply print the function object (without parentheses), i.e. \code{maxima.options} or \code{print(maxima.options)}.
}
