% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{maxima.repl}
\alias{maxima.repl}
\title{Run a Maxima REPL}
\usage{
maxima.repl(history_filename = NA_character_)
}
\arguments{
\item{history_filename}{A (optional) filename to which all Maxima commands
during the repl session are saved. If not provided (default), command history
lookup will still be available (if supported by OS) in the same way as in the
R session.}
}
\description{
This function provides a Maxima REPL in the \R session, which can be used 
to interactively run Maxima code. All code executed within the REPL is
run within the interactive Maxima session that is started when rim is 
attached and any generated Maxima objects will persist in the Maxima
session after the REPL is detached.
}
\details{
When working with R and Maxima scripts interactively, one can activate
the Python REPL with `maxima.repl()`, run Maxima code, and later run `exit;`
to return to the \R console.

Note that, inside the REPL, multiple commands are allowed.

The output format displayed inside the REPL is controlled by `maxima.options(repl.format = ...)`.
}
