% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.stop.R
\name{maxima.stop}
\alias{maxima.stop}
\title{Stoping maxima process}
\usage{
maxima.stop()
}
\value{
No return value
}
\description{
Calling this function stops the maxima child process and deletes the corresponding objects
}
\seealso{
Other rmaxima functions: 
\code{\link{maxima.start}()}
}
\concept{rmaxima functions}
