% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.tell.R
\name{maxima.tell}
\alias{maxima.tell}
\title{Execute maxima command and retrieve result}
\usage{
maxima.tell(command)
}
\arguments{
\item{command}{A character vector containing the maxima command.}
}
\value{
A character vector containing the corresponding output from maxima
}
\description{
Execute maxima command and retrieve result
}
\examples{
m <- new(RMaxima)

maxima.tell("2+2")
maxima.tell("2+2;")
maxima.tell("jacobian( [alpha / (alpha + beta), 1 / sqrt(alpha + beta)], [alpha, beta] )")
}
