% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.utils.R
\name{rim_global}
\alias{rim_global}
\title{rim_global}
\usage{
rim_global()
}
\value{
An environment, either the result of knitr::knit_global or maxima.env
}
\description{
Returns knitr::knit_global() unless this returs the GlobalEnv() in which case it returns maxima.env, rim's internal environment
}
