% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaxima-package.R
\docType{package}
\name{rmaxima-package}
\alias{rmaxima}
\alias{rmaxima-package}
\alias{maxima.start}
\alias{maxima.stop}
\alias{maxima.get}
\alias{maxima.load}
\alias{maxima.apropos}
\alias{maxima.setformat}
\alias{maxima.getformat}
\title{rmaxima}
\usage{
maxima.start(restart = FALSE)

maxima.stop()

maxima.get(command)

maxima.load(module)

maxima.apropos(keystring)

maxima.setformat(format = "linear")

maxima.getformat()
}
\arguments{
\item{restart}{If FALSE (default), then maxima process is started provided it is not running already. If TRUE starts or restarts maxima.}

\item{command}{A character vector containing the maxima command.}

\item{module}{A character vector naming the maxima module to be loaded.}

\item{format}{A character vector naming the output display format. Can be one of \code{linear} (default), \code{latex} (i.e. $$...$$), \code{text}.}
}
\description{
This the description bla bla bla
}
\details{
Here comes the details section
}
\section{Functions}{
\itemize{
\item \code{maxima.stop}: Quits the running maxima process.

\item \code{maxima.get}: Executes a single Maxima command provided by \code{command}.

\item \code{maxima.load}: A wrapper function to load a Maxima module named by \code{module}

\item \code{maxima.apropos}: A wrapper function to call the Maxima helper function \code{apropos} on a search term provided by \code{keystring}

\item \code{maxima.setformat}: Sets the format of the output string from Maxima.

\item \code{maxima.getformat}: Returns the currently set format as a character vector
}}

\author{
\strong{Maintainer}: Eric Stemmler \email{stemmler.eric@gmail.com}

Authors:
\itemize{
  \item Kseniia Shumelchyk \email{shumelchyk@gmail.com}
  \item Hans W. Borchers \email{hwborchers@googlemail.com}
}

}
\keyword{internal}
