% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{RTM}
\alias{RTM}
\title{Turn R data to Maxima Format}
\usage{
RTM(x, maximaName = NULL)
}
\arguments{
\item{x}{A object.}

\item{maximaName}{x object name.}
}
\value{
A string.
}
\description{
Turn R data to Maxima Format
}
\examples{
RTM(matrix(1), "henlo")
}
