% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{maxima.options}
\alias{maxima.options}
\title{maxima.options}
\usage{
maxima.options(
  ...,
  RESET = FALSE,
  READ.ONLY = NULL,
  LOCAL = FALSE,
  ADD = FALSE
)
}
\arguments{
\item{format}{Character vector of length 1.}

\item{engine.format}{...}

\item{inline.format}{...}

\item{label}{...}
}
\description{
Function for globally setting various options.
}
