% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima.start.R
\name{maxima.start}
\alias{maxima.start}
\title{Starting maxima}
\usage{
maxima.start(restart = FALSE)
}
\value{
No return value
}
\description{
Calling this function starts maxima in a child process in the background and configures the display output of maxima
}
\seealso{
Other rmaxima functions: 
\code{\link{maxima.stop}()}
}
\concept{rmaxima functions}
